django/requirements.txt
asgiref==3.7.2
attrs==23.2.0
beautifulsoup4==4.12.3
certifi==2024.2.2
charset-normalizer==3.3.2
colorclass==2.2.2
dj-database-url==2.1.0
Django==4.2.11
django-bootstrap-v5==1.0.11
django-cors-headers==4.3.1
django-djconfig==0.11.0
django-environ==0.11.2
django-filter==23.5
django-haystack==3.2.1
django-infinite-scroll-pagination==1.2.1
django-widget-tweaks==1.5.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
docopt==0.6.2
drf-spectacular==0.27.1
gunicorn==21.2.0
idna==3.6
inflection==0.5.1
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
mistune==3.0.2
olefile==0.47
packaging==23.2
pillow==10.2.0
pip-upgrader==1.4.15
psycopg2-binary==2.9.9
PyJWT==2.8.0
pytz==2023.4
PyYAML==6.0.1
referencing==0.33.0
requests==2.31.0
rest-framework-simplejwt==0.0.2
rpds-py==0.18.0
soupsieve==2.5
sqlparse==0.4.4
terminaltables==3.1.10
typing_extensions==4.10.0
tzdata==2024.1
uritemplate==4.1.1
urllib3==2.2.1
whitenoise==6.6.0
Whoosh==2.7.4
